Completed switchTitle Function
This is now working well now, for a single place to switch titles in a
page. It works well in both random and non-random modes.  Tested it with
a second TitleSwitcher implementation on the same page with only three
titles -- it appears to work for the first four titles, then it just
stops working. No error is output, so I will have to debug further.

I did not implement the click to switch title function, maybe I will add
this in later. Either way, the current format is much improved.